+++ games/game4/HeartsGameData.py
@@ -2,4 +2,6 @@
 
 class HeartsGameData:
     def __init__(self):
-        pass
+        self.cards = dict()
+        self.scores = dict()
+        self.starter_player = None